// web/src/api/generated/models/PartnerView.ts
/* tslint:disable */
/* eslint-disable */
/**
 * eimir.
 * Application Core. OpenAPI is the authoritative contract.
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * Account projection exposed through a space response.
 * 
 * This is deliberately an allowlist. Accounts also contain authentication
 * and contact data that do not belong in a space response; serializing the
 * general model would eventually expose such fields by accident.
 * @export
 * @interface PartnerView
 */
export interface PartnerView {
    /**
     * 
     * @type {string}
     * @memberof PartnerView
     */
    displayName: string;
    /**
     * 
     * @type {string}
     * @memberof PartnerView
     */
    id: string;
}

/**
 * Check if a given object implements the PartnerView interface.
 */
export function instanceOfPartnerView(value: object): value is PartnerView {
    if (!('displayName' in value) || value['displayName'] === undefined) return false;
    if (!('id' in value) || value['id'] === undefined) return false;
    return true;
}

export function PartnerViewFromJSON(json: any): PartnerView {
    return PartnerViewFromJSONTyped(json, false);
}

export function PartnerViewFromJSONTyped(json: any, ignoreDiscriminator: boolean): PartnerView {
    if (json == null) {
        return json;
    }
    return {
        
        'displayName': json['displayName'],
        'id': json['id'],
    };
}

export function PartnerViewToJSON(json: any): PartnerView {
    return PartnerViewToJSONTyped(json, false);
}

export function PartnerViewToJSONTyped(value?: PartnerView | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'displayName': value['displayName'],
        'id': value['id'],
    };
}